Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main()
 {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int total=3;
-    printf("Average: %d\n",a+b+c/total);
+    float total=3.0;
+    printf("Average: %.2f\n",a+b+c/total);
     return 0;
 }
